Fix part of a problem rendering date objects
It fixes temporally a problem with electron passing
date objects through the remote.
It is already fixed in the latest electron version
but we still can't use it because the latest have another
problem passing exceptions though the remote.

This way we can see the date value. But at least we can see the rest of
the result.