Improved performance during the rendering (close #61)
Added pagination for the query result. This way is necessary just render
the current page. Which improves a lot the performance. I got 10 secs
of difference between the same query before and after my changes.

I also had to include in the same commit a fix to the “copy action”. It
sets the copied attribute only after the copying has finished. Also removed
an workaround for race conditions in the state change.