Do not pass the main window to menu builder
There is no need for that. Since the current win is available in the
click arguments. It fixes a problem handling IPC messages on running 
multiple windows.

But with this fixing was found another bug related. The app does not handle
multiple connections over different windows. The same connection is shared over
every widow. Lets fix it in other commit.