Change close connection button to actually close the connection
The previous behavior was just closing the query browser window.
But was keeping the connection opened for cache reasons for a next connection.
Although it was only useful if you would connect to the same server.
Which is not so frequent to happen. Usually if I a leave a server I will
connect to another in next connection.