Improve state handling between connections
Previously each new connection would unmount almost all components in the whole page.
Which as causing the components to lose their state.

It now will only clean the whole store in server connections (the first connection to a server).

I also broke down the database components in three pieces to be easier handle the data/state for each one of them.