Change short keys to focus into the input fields
These short keys seems a bit eaiser to rember.
Database and DB meta objects that stay in the left of the screen use "(" as base key.
And the query box that stay in the right of the screen use ")" as base key.

I tried to use "<" and ">" but the "<" key shares the same key for "," which is already used by ace editor. Even not using that field I couldn't disable that shortkey on ace editor.