Show more error context
It will show all the properties available in the error object.
In same cases it really is more useful than only the error message.
For example on PG on creating a unique index over a duplicated value it will show the duplicated value.
For other cases it still quite useful as well because will should the line of the error.

Close #209.