update compile time dependencies (#488)
This updates "compile time" dependencies, focusing principally
on babel, eslint, and webpack. Upgrading these however includes
three large-ish BC breaks, namely Babel 6 -> 7 and webpack 1 -> 2/3,
and then 2/3 -> 4. While it would have probably been nice to have
upgraded them one at a time, each of these dependencies are so
intertwined that it's easier to upgrade all at once and be done with
it.

As part of the upgrade of eslint, I did fix some of the new lint errors, but 
I added a number of ignore rules that can be solved over time.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>