remove unused react-transform-hmr
Usage of this module was removed in the upgrade to a newer version
of babel / webpack. It's long since been deprecated anyway, and
it is now suggested to use react-hot-loader instead, except that
might also go away soon in favor of React AutoRefresh, though that
is not yet fully stable, so not yet putting in the effort to bring
it back. In the meantime, the system is already set-up such that
making changes to any file will cause a restart of electron through
the spawn-auto-restart at this time.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>